+++ nodebackend/src/utils/buildUrlMap.ts
@@ -1,7 +1,6 @@
-import { connectToMongo } from '../clients/mongoClient.js';
-import { ObjectId } from 'mongodb';  
 import * as envSwitcher from '../envSwitcher.js';
 import logger from '../logger.js';
+import { irrigationSwitchTopics, irrigationTuyaDatapoints } from './constants.js';
 
 const baseUrl: string = envSwitcher.baseUrl;
 
@@ -9,31 +8,14 @@ interface UrlMap {
     [key: string]: string;
 }
 
-interface Document {
-    mqttTopics: string[];
-    vAtuyaUrl: string[];
-}
-
 async function buildUrlMap(): Promise<UrlMap> {
     try {
-        const db = await connectToMongo();
-
-        const collection = db.collection('nodeServerConfig');
-        const docId: string = "652e8f1a49124f743556be68";
-
-        // Use type assertion here
-        const doc = await collection.findOne({ _id: new ObjectId(docId) }) as Document | null;
-
-        if (!doc) {
-            throw new Error(`Document with id ${docId} not found.`);
-        }
-
         const urlMap: UrlMap = {};
 
-        for (let i = 0; i < doc.mqttTopics.length; i++) {
-            const topic: string = doc.mqttTopics[i];
-            const url: string = doc.vAtuyaUrl[i];
-            urlMap[topic] = `${baseUrl}/set/${url}`;
+        for (let i = 0; i < irrigationSwitchTopics.length; i++) {
+            const topic: string = irrigationSwitchTopics[i];
+            const dp: string = irrigationTuyaDatapoints[i];
+            urlMap[topic] = `${baseUrl}/set/${dp}`;
         }
         return urlMap;
     } catch (error) {